+++ genericsuite_ai/lib/ibm.py
@@ -8,13 +8,12 @@ https://developer.ibm.com/
 import requests
 
 from typing import Any, List, Optional
-# from typing import Mapping, Dict, Iterator
 
 from langchain_core.callbacks.manager import CallbackManagerForLLMRun
 # from langchain_core.outputs import GenerationChunk
 
-# from langchain_core.language_models.llms import LLM
 from genericsuite_ai.lib.ai_langchain_models_abstract import CustomLLM as LLM
+from genericsuite.util.app_context import AppContext
 from genericsuite.util.app_logger import log_debug
 
 
@@ -31,6 +30,7 @@ class IbmWatsonx(LLM):
     project_id: str
     model_url: str
     identity_token_url: str
+    app_context: Optional[AppContext] = None
 
     def _call(
         self,